+++ publish/CHANGELOG.md
@@ -3,6 +3,7 @@
   - Adds a new field `clearRadius` to the `expand_vegetation.yaml` file.
   - Adds new fields `statusEffects`, `dayStatusEffects` and `nightStatusEffects` to the `expand_environments.yaml` file.
   - Changes the fields `objectSwap` and `objectData` in the `expand_locations.yaml` to also affect dungeons (reverting previous change).
+  - Disables minimap generation if Better Continents is enabled (performance + compatiblity).
   - Fixes the scale not working on extra objects.
   - Fixes dungeon being blocked by CLLC dungeon reset.